requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pytesseract==0.3.10
pdfplumber==0.9.0
chromadb==0.3.27
pdf2image
camelot 
ghostscript
google-generativeai==0.8.3
PyPDF2==3.0.1
numpy==1.24.4
pillow==9.4.0







